Added indentation for example & small edit in TODO title
There is weird issue with Nim and multi-line blocks (in that case, the example). Normally, you can do it as I did (with a one-tab indentation), so that it looks good, legible and pretty much like all other examples. The only difference being when the function itself is in a e.g. `when defined(WEB)` block, where there should be no indentation. (Basically, that's a Nim bug...)

That aside, great job!